requirements.txt
Pillow==10.1.0
python-barcode==0.15.1